cleaned up some warnings and removed cout from constructors

+++ CSD2b/eindopdracht_SynthSong/src/bitcrusher.h
@@ -10,7 +10,7 @@ class BitCrusher {
 
     // bitreduction
     float applyBitReduction(float sample);
-    void setBitDepth(int bitDepth);
+    void setBitDepth(float bitDepth);
     void setBypassState(bool bypass);
     bool getBypassState();
